Harden Apple governance mutation receipts
